rag_app/requirements.txt
langchain
langchain-community
python-dotenv
langsmith
unstructured
docx2txt
python-docx
reportlab
chromadb
streamlit
